# [GPTNeo] create local attention mask ones